+++ source/common/config/filesystem_subscription_impl.h
@@ -6,12 +6,6 @@
 #include "envoy/event/dispatcher.h"
 #include "envoy/filesystem/filesystem.h"
 
-#include "common/common/logger.h"
-#include "common/common/macros.h"
-#include "common/config/utility.h"
-#include "common/protobuf/protobuf.h"
-#include "common/protobuf/utility.h"
-
 namespace Envoy {
 namespace Config {
 
@@ -24,59 +18,16 @@ class FilesystemSubscriptionImpl : public Config::Subscription,
                                    Logger::Loggable<Logger::Id::config> {
 public:
   FilesystemSubscriptionImpl(Event::Dispatcher& dispatcher, absl::string_view path,
-                             SubscriptionStats stats, Api::Api& api)
-      : path_(path), watcher_(dispatcher.createFilesystemWatcher()), stats_(stats), api_(api) {
-    watcher_->addWatch(path_, Filesystem::Watcher::Events::MovedTo, [this](uint32_t events) {
-      UNREFERENCED_PARAMETER(events);
-      if (started_) {
-        refresh();
-      }
-    });
-  }
+                             SubscriptionStats stats, Api::Api& api);
 
   // Config::Subscription
   void start(const std::set<std::string>& resources,
-             Config::SubscriptionCallbacks& callbacks) override {
-    // We report all discovered resources in the watched file.
-    UNREFERENCED_PARAMETER(resources);
-    callbacks_ = &callbacks;
-    started_ = true;
-    // Attempt to read in case there is a file there already.
-    refresh();
-  }
+             Config::SubscriptionCallbacks& callbacks) override;
 
-  void updateResources(const std::set<std::string>& resources) override {
-    // We report all discovered resources in the watched file.
-    UNREFERENCED_PARAMETER(resources);
-    // Bump stats for consistence behavior with other xDS.
-    stats_.update_attempt_.inc();
-  }
+  void updateResources(const std::set<std::string>& resources) override;
 
 private:
-  void refresh() {
-    ENVOY_LOG(debug, "Filesystem config refresh for {}", path_);
-    stats_.update_attempt_.inc();
-    bool config_update_available = false;
-    try {
-      envoy::api::v2::DiscoveryResponse message;
-      MessageUtil::loadFromFile(path_, message, api_);
-      config_update_available = true;
-      callbacks_->onConfigUpdate(message.resources(), message.version_info());
-      stats_.version_.set(HashUtil::xxHash64(message.version_info()));
-      stats_.update_success_.inc();
-      ENVOY_LOG(debug, "Filesystem config update accepted for {}: {}", path_,
-                message.DebugString());
-    } catch (const EnvoyException& e) {
-      if (config_update_available) {
-        ENVOY_LOG(warn, "Filesystem config update rejected: {}", e.what());
-        stats_.update_rejected_.inc();
-      } else {
-        ENVOY_LOG(warn, "Filesystem config update failure: {}", e.what());
-        stats_.update_failure_.inc();
-      }
-      callbacks_->onConfigUpdateFailed(&e);
-    }
-  }
+  void refresh();
 
   bool started_{};
   const std::string path_;
